Python_introduction

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 income = float(input("Enter your monthly income: "))
 expense = float(input("Enter your total monthly expenses: "))
-monthly_saving = income - expense
-projected_saving = monthly_saving * 12 + (monthly_saving * 12 * 0.05)
+monthly_saving = int(income - expense)
+projected_saving = int(monthly_saving * 12 + (monthly_saving * 12 * 0.05))
 print("Your monthly savings are $",monthly_saving)
 print("Projected savings after one year, with interest, is : $",projected_saving)
